% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom.R
\name{mom}
\alias{mom}
\title{An xtsmatt Function}
\usage{
mom(x, merge = FALSE)
}
\arguments{
\item{x}{The name of your xts object.}

\item{merge}{If TRUE, merges orginal xts object with newly calculated object. If FALSE, returns just the calculated matrix.}
}
\description{
This function allows you to calculate month over month % change for an xts object.
}
\examples{
mom(myxts_obj, merge = TRUE)


}
\keyword{mom,}
\keyword{month}
\keyword{over}
