% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prv_mon.R
\name{prv_mon}
\alias{prv_mon}
\title{An xtsmatt Function}
\usage{
prv_mon(x, merge = FALSE)
}
\arguments{
\item{x}{The name of your xts object.}

\item{merge}{If TRUE, merges orginal xts object with newly calculated object. If FALSE, returns just the calculated matrix.}
}
\description{
Calculate % change from the immediatly prior month.
}
\examples{
prv_mon(myxts_obj, TRUE)


}
\keyword{change,}
\keyword{delta}
\keyword{month,}
\keyword{percent,}
\keyword{prv_mon,}
