% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prv_qtr.R
\name{prv_qtr}
\alias{prv_qtr}
\title{An xtsmatt Function}
\usage{
prv_qtr(x, merge = FALSE)
}
\arguments{
\item{x}{The name of your xts object.}

\item{merge}{If TRUE, merges orginal xts object with newly calculated object. If FALSE, returns just the calculated matrix.}
}
\description{
Calculate % change from the immediatly prior qtr.
}
\examples{
prv_qtr(myxts_obj, TRUE)


}
\keyword{change,}
\keyword{delta}
\keyword{percent,}
\keyword{prv_qtr,}
\keyword{quarter,}
