% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prv_wk.R
\name{prv_wk}
\alias{prv_wk}
\title{An xtsmatt Function}
\usage{
prv_wk(x, merge = FALSE)
}
\arguments{
\item{x}{The name of your xts object.}

\item{merge}{If TRUE, merges orginal xts object with newly calculated object. If FALSE, returns just the calculated matrix.}
}
\description{
Calculate % change from the immediatly prior wk.
}
\examples{
prv_wk(myxts_obj, TRUE)


}
\keyword{change,}
\keyword{delta}
\keyword{percent,}
\keyword{prv_wk,}
\keyword{quarter,}
