% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qoq.R
\name{qoq}
\alias{qoq}
\title{An xtsmatt Function}
\usage{
qoq(x, merge = FALSE)
}
\arguments{
\item{x}{The name of your xts object.}

\item{merge}{If TRUE, merges orginal xts object with newly calculated object. If FALSE, returns just the calculated matrix.}
}
\description{
This function allows you to calculate quarter over quarter % change for an xts object.
}
\examples{
qoq(myxts_obj, TRUE)


}
\keyword{qoq,}
\keyword{quarter}
