% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wow.R
\name{wow}
\alias{wow}
\title{An xtsmatt Function}
\usage{
wow(x, merge = FALSE)
}
\arguments{
\item{x}{The name of your xts object.}

\item{merge}{If TRUE, merges orginal xts object with newly calculated object. If FALSE, returns just the calculated matrix.}
}
\description{
This function allows you to calculate week over week % change from the previous year for an xts object.
}
\examples{
wow(myxts_obj, merge = TRUE)


}
\keyword{over}
\keyword{week}
\keyword{wow,}
