% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yoy.R
\name{yoy}
\alias{yoy}
\title{An xtsmatt Function}
\usage{
yoy(x, merge = FALSE)
}
\arguments{
\item{x}{The name of your xts object.}

\item{merge}{If TRUE, merges orginal xts object with newly calculated object. If FALSE, returns just the calculated matrix.}
}
\description{
This function allows you to calculate year over year % change for an xts object.
}
\examples{
yoy(myxts_obj, merge = TRUE)


}
\keyword{over}
\keyword{year}
\keyword{yoy,}
